agent/internal: rename package names into internal
