+++ agent/internal/metrics.go
@@ -1,4 +1,4 @@
-package agent
+package internal
 
 import (
 	"context"
@@ -7,8 +7,8 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/sqreen/go-agent/agent/backend/api"
-	"github.com/sqreen/go-agent/agent/config"
+	"github.com/sqreen/go-agent/agent/internal/backend/api"
+	"github.com/sqreen/go-agent/agent/internal/config"
 )
 
 type metricsManager struct {
