+++ agent/internal/version.go
@@ -1,3 +1,3 @@
-package agent
+package internal
 
 const version = "0.1.0-beta.2"
